Moved strings to a unified folder
