+++ src/strings/editor.ts
