+++ src/strings/syntax-nodes.ts
